pyproject.toml
[tool.poetry]
name = "fastapi-graphql"
version = "0.1.0"
description = ""
authors = ["zhanchao.wu <zhanchao.wu@icloud.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.101.1"
uvicorn = { extras = ["standard"], version = "^0.23.2" }
pydantic = "^2.2.1"
tortoise-orm = { extras = ["asyncmy"], version = "^0.20.0" }
aerich = "^0.7.2"
python-dotenv = "^1.0.0"
pydantic-settings = "^2.0.3"
strawberry-graphql = {extras = ["fastapi"], version = "^0.204.0"}
cryptography = "^41.0.3"
pygments = "^2.16.1"


[tool.poetry.group.dev.dependencies]
black = "*"
ruff = "*"
pre-commit = "^3.3.3"
commitizen = "^3.6.0"


[tool.aerich]
tortoise_orm = "fastapi_graphql.config.env_setting.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
select = [
    "E",   # pycodestyle error
    "W",   # pycodestyle warning
    "F",   # pyflakes
    "A",   # flakes8-builtins
    "COM", # flakes8-commas
    "C4",  # flake8-comprehensions
    "Q",   # flake8-quotes
    "SIM", # flake8-simplify
    "PTH", # flake8-use-pathlib
    #   "I",   # isort
    "N",  # pep8 naming
    "UP", # pyupgrade  
    "S",  # bandit
]

[tool.commitizen]
name = "cz_conventional_commits"
version = "1.0.2"
tag_format = "$version"
